精简代码
