+++ Masuit.Tools.Core/Systems/StopwatchHelper.cs
@@ -3,16 +3,8 @@ using System.Diagnostics;
 
 namespace Masuit.Tools.Systems
 {
-    /// <summary>
-    /// stopwatch扩展
-    /// </summary>
-    public static class StopwatchExtension
+    public static class StopwatchHelper
     {
-        /// <summary>
-        /// 检测方法执行时间
-        /// </summary>
-        /// <param name="action"></param>
-        /// <returns></returns>
         public static double Execute(Action action)
         {
             Stopwatch sw = Stopwatch.StartNew();
