+++ Masuit.Tools/Net/CookieHelper.cs
@@ -1,12 +1,4 @@
-﻿using Masuit.Tools.Logging;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Net;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
+﻿using System;
 using System.Web;
 
 namespace Masuit.Tools.Net
@@ -16,66 +8,11 @@ namespace Masuit.Tools.Net
     /// </summary>
     public static class CookieHelper
     {
-        [DllImport("wininet.dll", SetLastError = true)]
-        private static extern bool InternetGetCookie(string url, string cookieName, StringBuilder cookieData, ref int size);
-
-        /// <summary>
-        /// 获取Cookie容器
-        /// </summary>
-        /// <param name="uri">URL</param>
-        /// <returns>Cookie容器</returns>
-        public static CookieContainer GetUriCookieContainer(this Uri uri)
-        {
-            CookieContainer cookies = null;
-            //定义Cookie数据的大小。   
-            int datasize = 256;
-            StringBuilder cookieData = new StringBuilder(datasize);
-            if (!InternetGetCookie(uri.ToString(), null, cookieData, ref datasize))
-            {
-                if (datasize < 0) return null;
-                // 确信有足够大的空间来容纳Cookie数据。   
-                cookieData = new StringBuilder(datasize);
-                if (!InternetGetCookie(uri.ToString(), null, cookieData, ref datasize)) return null;
-            }
-
-            if (cookieData.Length <= 0)
-            {
-                return cookies;
-            }
-
-            cookies = new CookieContainer();
-            cookies.SetCookies(uri, cookieData.ToString().Replace(';', ','));
-
-            return cookies;
-        }
-
-        /// <summary>
-        /// 获取url下的Cookie
-        /// </summary>
-        /// <param name="cookies">Cookie对象</param>
-        /// <param name="uri">URL</param>
-        /// <returns>Cookie内容</returns>
-        public static string PrintCookies(this CookieContainer cookies, Uri uri)
-        {
-            CookieCollection cc = cookies.GetCookies(uri);
-            StringBuilder sb = new StringBuilder();
-            foreach (Cookie cook in cc)
-            {
-                sb.AppendLine(string.Format("{0}:{1}", cook.Name, cook.Value));
-            }
-
-            return sb.ToString();
-        }
-
-        #region 清除CooKie 
-
-        #region 清除指定Cookie
-
         /// <summary>
         /// 清除指定Cookie
         /// </summary>
         /// <param name="cookiename">cookiename</param>
-        public static void ClearCookie(string cookiename)
+        public static void Clear(string cookiename)
         {
             HttpCookie cookie = HttpContext.Current.Request.Cookies[cookiename];
             if (cookie != null)
@@ -85,14 +22,10 @@ namespace Masuit.Tools.Net
             }
         }
 
-        #endregion
-
-        #region   删除所有cookie值
-
         /// <summary>
         /// 删除所有cookie值
         /// </summary>
-        public static void Clear()
+        public static void ClearAll()
         {
             int n = HttpContext.Current.Response.Cookies.Count;
             for (int i = 0; i < n; i++)
@@ -102,60 +35,6 @@ namespace Masuit.Tools.Net
                 HttpContext.Current.Response.Cookies.Add(myCookie);
             }
         }
-
-        #endregion
-
-        #region 清除系统指定的Cookie
-
-        /// <summary>
-        /// 清除系统指定的Cookie
-        /// </summary>
-        /// <param name="ck">指定的Cookie</param>
-        /// <param name="url">Cookie的域</param>
-        public static void ClearCookies(string ck, string url)
-        {
-            string[] cks = GetCKS(ck);
-            List<string> domains = GetDomains(url);
-            for (int i = 0; i < cks.Length; i++)
-            {
-                string[] nv = cks[i].Split('=');
-                for (int j = 0; j < domains.Count; j++)
-                {
-                    InternetSetCookie(domains[j], nv[0], "abc;expires = Sat, 31-Dec-2007 14:00:00 GMT");
-                }
-            }
-        }
-
-        #endregion
-
-        #region 清除系统的Cookie文件
-
-        /// <summary>
-        /// 清除系统的Cookie文件
-        /// </summary>
-        public static void ClearCookiesFiles()
-        {
-            DirectoryInfo di = new DirectoryInfo(Environment.GetFolderPath(Environment.SpecialFolder.Cookies));
-            foreach (FileInfo file in di.GetFiles())
-            {
-                try
-                {
-                    file.Delete();
-                }
-                catch (Exception e)
-                {
-                    LogManager.Error(e);
-                }
-            }
-        }
-
-        #endregion
-
-        #endregion
-
-        #region 获取Cookie
-
-        #region 获取指定Cookie值
 
         /// <summary>
         /// 获取指定Cookie值
@@ -175,181 +54,6 @@ namespace Masuit.Tools.Net
         }
 
         /// <summary>
-        /// 根据Key值得到Cookie值,Key不区分大小写
-        /// </summary>
-        /// <param name="key">key</param>
-        /// <param name="cookie">字符串Cookie</param>
-        /// <returns>Cookie值</returns>
-        public static string GetCookieValue(string key, string cookie)
-        {
-            foreach (CookieItem item in GetCookieList(cookie))
-            {
-                if (item.Key == key)
-                    return item.Value;
-            }
-
-            return "";
-        }
-
-        /// <summary>
-        /// 根据字符生成Cookie列表
-        /// </summary>
-        /// <param name="cookie">Cookie字符串</param>
-        /// <returns>Cookie列表</returns>
-        public static List<CookieItem> GetCookieList(string cookie)
-        {
-            List<CookieItem> cookielist = new List<CookieItem>();
-            foreach (string item in cookie.Split(new[]
-            {
-                ";",
-                ","
-            }, StringSplitOptions.RemoveEmptyEntries))
-            {
-                if (Regex.IsMatch(item, @"([\s\S]*?)=([\s\S]*?)$"))
-                {
-                    Match m = Regex.Match(item, @"([\s\S]*?)=([\s\S]*?)$");
-                    cookielist.Add(new CookieItem()
-                    {
-                        Key = m.Groups[1].Value,
-                        Value = m.Groups[2].Value
-                    });
-                }
-            }
-
-            return cookielist;
-        }
-
-        #endregion
-
-        #region  获取cookie数组
-
-        /// <summary>
-        /// 获取cookie数组
-        /// </summary>
-        /// <param name="ck">cokki数据</param>
-        /// <returns>字符串数组</returns>
-        public static string[] GetCKS(string ck)
-        {
-            return ck == null ? new string[0] : ck.Split(";".ToCharArray(), StringSplitOptions.RemoveEmptyEntries);
-        }
-
-        #endregion
-
-        #region  从Cookie数组中转换成不重复的Cookie字符串，相同的Cookie取前面的
-
-        /// <summary>
-        /// 从Cookie数组中转换成不重复的Cookie字符串，相同的Cookie取前面的
-        /// </summary>
-        /// <param name="cks">Cookie数组</param>
-        /// <returns>剔除重复的Cookie字符串</returns>
-        public static string GetCK(string[] cks)
-        {
-            string res = "";
-            List<string> list = new List<string>();
-            for (int i = 0; i < cks.Length; i++)
-            {
-                if (cks[i].Trim() != "")
-                {
-                    if (!IncludeCKAsync(list, cks[i]).Result)
-                    {
-                        list.Add(cks[i].Trim());
-                    }
-                }
-            }
-
-            foreach (var t in list)
-            {
-                res += t + ";";
-            }
-
-            return res;
-        }
-
-        #endregion
-
-        #region  从CookieCollection中获取Cookie字符串
-
-        /// <summary>
-        /// 从CookieCollection中获取Cookie字符串
-        /// </summary>
-        /// <param name="cc">CookieCollection，一般用在WebRequest中</param>
-        /// <returns>转换后的Cookie字符串</returns>
-        public static string GetCK(this CookieCollection cc)
-        {
-            string ck = "";
-            for (int i = 0; i < cc.Count; i++)
-            {
-                ck += cc[i].Name + "=" + cc[i].Value + ";";
-            }
-
-            return ck;
-        }
-
-        #endregion
-
-        #region   将Cookie字符串填充到CookieCollection中
-
-        /// <summary>
-        /// 将Cookie字符串填充到CookieCollection中
-        /// </summary>
-        /// <param name="ck">Cookie字符串</param>
-        /// <param name="url">Cookie的域</param>
-        /// <returns>添加后的CookieCollection</returns>
-        public static CookieCollection GetCK(string ck, string url)
-        {
-            CookieCollection cc = new CookieCollection();
-            Uri u = new Uri(url);
-            var domain = u.Host;
-            string[] cks = GetCKS(ck);
-            foreach (var t in cks)
-            {
-                if (t.IndexOf("=") > -1)
-                {
-                    try
-                    {
-                        string n = t.Substring(0, t.IndexOf("="));
-                        string v = t.Substring(t.IndexOf("=") + 1);
-                        Cookie c = new Cookie
-                        {
-                            Name = n.Trim(),
-                            Value = v.Trim(),
-                            Domain = domain
-                        };
-                        cc.Add(c);
-                    }
-                    catch (Exception e)
-                    {
-                        LogManager.Error(e);
-                    }
-                }
-            }
-
-            return cc;
-        }
-
-        #endregion
-
-        #endregion
-
-        #region 添加Cookie
-
-        #region   添加一个Cookie（24小时过期）
-
-        /// <summary>
-        /// 添加一个Cookie（24小时过期）
-        /// </summary>
-        /// <param name="cookiename">键</param>
-        /// <param name="cookievalue">值</param>
-        public static void SetCookie(string cookiename, string cookievalue)
-        {
-            SetCookie(cookiename, cookievalue, DateTime.Now.AddDays(1.0));
-        }
-
-        #endregion
-
-        #region   添加一个Cookie
-
-        /// <summary>
         /// 添加一个Cookie
         /// </summary>
         /// <param name="cookiename">cookie名</param>
@@ -364,159 +68,5 @@ namespace Masuit.Tools.Net
             };
             HttpContext.Current.Response.Cookies.Add(cookie);
         }
-
-        #endregion
-
-        #endregion
-
-        #region 检查Cookie集合中是否包含指定的Cookie值
-
-        /// <summary>
-        /// 检查Cookie集合中是否包含指定的Cookie值
-        /// </summary>
-        /// <param name="cks">Cookie集合</param>
-        /// <param name="ck">待判断的Cookie</param>
-        /// <returns>Cookie集合中是否包含指定的Cookie</returns>
-        public static async Task<bool> IncludeCKAsync(this List<string> cks, string ck)
-        {
-            return await Task.Run(() =>
-            {
-                foreach (string t in cks)
-                {
-                    if (t.ToLower().Trim().IndexOf(ck.Remove(ck.IndexOf('=') + 1).Trim().ToLower(), StringComparison.Ordinal) != -1)
-                    {
-                        return true;
-                    }
-                }
-
-                return false;
-            }).ConfigureAwait(false);
-        }
-
-        #endregion
-
-        #region 设置系统Cookie
-
-        /// <summary>
-        /// 设置系统Cookie
-        /// </summary>
-        /// <param name="lpszUrlName">Cookie域</param>
-        /// <param name="lbszCookieName">Cookie名</param>
-        /// <param name="lpszCookieData">Cookie数据</param>
-        /// <returns>设置成功与否</returns>
-        [DllImport("wininet.dll", CharSet = CharSet.Auto, SetLastError = true)]
-        public static extern bool InternetSetCookie(string lpszUrlName, string lbszCookieName, string lpszCookieData);
-
-        #endregion
-
-        #region   获取所有可能的Cookie域
-
-        /// <summary>
-        /// 获取所有可能的Cookie域
-        /// </summary>
-        /// <param name="url">域的全称</param>
-        /// <returns>所有可能的域</returns>
-        public static List<string> GetDomains(string url)
-        {
-            List<string> res = new List<string>();
-            url = url.Remove(url.IndexOf("?"));
-            Uri uri = new Uri(url);
-            string baseDomain = uri.Scheme + "://" + uri.Host;
-            foreach (var t in uri.Segments)
-            {
-                baseDomain += t;
-                res.Add(baseDomain);
-            }
-
-            return res;
-        }
-
-        #endregion
-
-        #region  将Cookie字符串描述的Cookie追加到CookieCoollection
-
-        /// <summary>
-        /// 将Cookie字符串描述的Cookie追加到CookieCoollection
-        /// </summary>
-        /// <param name="cc">CookieCoollection</param>
-        /// <param name="ck">Cookie字符串</param>
-        /// <param name="url">Cookie的域</param>
-        public static void SetCKAppendToCC(this CookieCollection cc, string ck, string url)
-        {
-            CookieCollection tmp = GetCK(ck, url);
-            for (int i = 0; i < tmp.Count; i++)
-            {
-                cc.Add(tmp[i]);
-            }
-        }
-
-        #endregion
-
-        #region 将Cookie字符串设置到系统中，便于浏览器使用
-
-        /// <summary>
-        /// 将Cookie字符串设置到系统中，便于浏览器使用
-        /// </summary>
-        /// <param name="ck">Cookie字符串</param>
-        /// <param name="url">Cookie的域</param>
-        public static void SetCKToSystem(string ck, string url)
-        {
-            string[] cks = GetCKS(ck);
-            foreach (var t in cks)
-            {
-                string[] nv = t.Split('=');
-                InternetSetCookie(url, nv[0], nv.Length > 1 ? nv[1] : "");
-            }
-        }
-
-        #endregion
-
-        #region 将CookieCollection中的Cookie设置到系统中，便于浏览器使用
-
-        /// <summary>
-        /// 将CookieCollection中的Cookie设置到系统中，便于浏览器使用
-        /// </summary>
-        /// <param name="cc">CookieCollection</param>
-        /// <param name="url">Cookie的域</param>
-        public static void SetCKToSystem(this CookieCollection cc, string url)
-        {
-            List<string> domains = GetDomains(url);
-            for (int i = 0; i < cc.Count; i++)
-            {
-                foreach (var t in domains)
-                {
-                    InternetSetCookie(t, cc[i].Name, cc[i].Value);
-                }
-            }
-        }
-
-        #endregion
-
-        /// <summary>
-        /// 格式化Cookie为标准格式
-        /// </summary>
-        /// <param name="key">Key值</param>
-        /// <param name="value">Value值</param>
-        /// <returns>格式化后的数据</returns>
-        public static string CookieFormat(string key, string value)
-        {
-            return string.Format("{0}={1};", key, value);
-        }
-    }
-
-    /// <summary>
-    /// Cookie对象
-    /// </summary>
-    public class CookieItem
-    {
-        /// <summary>
-        /// 键
-        /// </summary>
-        public string Key { get; set; }
-
-        /// <summary>
-        /// 值
-        /// </summary>
-        public string Value { get; set; }
     }
 }
